stm class: Make p_basic a soft dependency of stm_core
It appears that an async request_module() can be replaced with a modprobe
"soft dep" hint, which should have the same effect given that you're using
a proper version of modprobe that actually knows to look there.

Replace the request_module call for the "p_basic" protocol with a soft dep
hint in the STM core.

Signed-off-by: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Reported-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>