x86/cpu: Enable LASS (Linear Address Space Separation)
Enable LASS for every CPU by setting the CR4 bit.

LASS is incompatible with the legacy vsyscall feature because
users must access the vsyscall functions mapped in the range
0xffffffffff600000-0xffffffffff601000 which will trigger
a #GP fault when LASS is enabled.

In theory it would be possible to write a #GP fault handler
to emulate the old vsyscall behavior, but vsyscall has been
deprecated for some time, so this has not been done by this
patch series and lass will always override the legacy vsyscalls.

Signed-off-by: Yian Chen <yian.chen@intel.com>