ACPI: Add SVKL table support
For highly secure systems like (TDX), the storage volume will
typically be an encrypted volume. In that case, by design, the
Firmware will need to support quote generation and attestation
to be able to fetch a set of storage-volume key(s) from a
remote-key server during boot and pass the key to the OS. Also
by design, the key is stored in the memory, and the information
of the key is passed from Firmware to OS via an ACPI table. More
details about ACPI Storage Volume Key Label (SVKL) table can be
found in TDX spec titled "Guest-Host-Communication Interface (GHCI)
for Intel® Trust Domain Extensions (Intel® TDX)", sec 4.4.

Also, in order to pass the received storage keys to userspace,
define following set of IOCTLs.

/* Get the total number of keys */
   #define ACPI_SVKL_GET_KEY_COUNT _IOW('E', 0x01, __u32)
/* Get the key header (by index) */
   #define ACPI_SVKL_GET_KEY_INFO  _IOWR('E', 0x02, key_info_size)
/* Get the key data (by index) */
   #define ACPI_SVKL_GET_KEY_DATA  _IOR('E', 0x03, __u64)
/* Clear the key data (by index) */
   #define ACPI_SVKL_CLEAR_KEY     _IOR('E', 0x04, __u32)

Signed-off-by: Kuppuswamy Sathyanarayanan <sathyanarayanan.kuppuswamy@linux.intel.com>