x86/cpufeatures: Add feature bit for HRESET
HRESET (history reset) requests the processor to reset selected components
of hardware history maintained by he current logical processor.

An instance of such hardware history is that collected by the Intel Thread
Director (ITD). ITD provides on resources in hardware to build a history
of a software thread running on the logical processor. Such history is
then used to classify a software threads in two or more classes. Once
classified, ITD assigns a classID to the software thread.

Since multiple threads may run on the same logical processor, it is
necessary to reset the hardware history when switching context. The
instruction HRESET serves this purpose.

MSR_IA32_HW_HRESET_ENABLE is a per-CPU register whose bits selects the
hardware history components to reset when HRESET is executed.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>