sched/task_struct: Introduce classes of tasks
On hybrid processors, the architecture differences between the types of
CPUs lead to different instructions-per-cycle (IPC) on each type of CPU.
IPCs may differ further by the type of instructions. Instructions can be
grouped into classes of similar IPCs.

Hence, tasks may be classified into groups depending on the type of
instructions they execute.

Add a new member task_struct::classid to associate a particular task to
a classification ID that depends on the instructions it executes. Add the
extra management members classid_candidate and classid_debounce_counter
hardware implementations can use to compute a classid that the scheduler
consume.

The scheduler may use the classification ID of a task and information about
the relative performance among CPUs of a particular classification ID to
improve throughput. It may, for instance, place a given task on a CPU that
has a higher IPC.

The methods to determine the classification ID of a task and its relative
IPC are specific to each CPU archiecture.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>