x86/hreset: Configure history reset
Recently released processors such as Alderlake build a per-CPU history of
the software thread in execution. Ror instance, the Intel Thread Director
builds history to classify the software thread into any of the classes of
the Hardware Feedback Interface table).

It sometimes necessary to reset the history that hardware has built (for
insance, when switching context to a new software thread). History is
reset with the HRESET instruction.

Hardware may support more than one history components. The
MSR IA32_HW_RESET_ENABLE is a per-CPU that selects which history components
are reset when HRESET is executed on the CPU. The supported history
components are specified in CPUID(0x20).EBX.

Configure history reset on each CPU. Configure IA32_HW_RESET_ENABLE so
that all histories are reset when HRESET is executed. Use the history
components that the boot CPU supports, since all other secondary CPUs
support it.

A static global variable caches the history components that the platform
supports. This information will be needed when resetting history during
context switch.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>