thermal: intel: hfi: Update the class of the current task
Intel Thread Director classifies a running task based on the instructions
it executes. The classification that hardware provide points to a column
in the Hardware Feedback Interface table. The scheduler may use these two
pieces of information to determine on which CPU the task must be placed.

Implement the interfaces that the scheduler define to update the
classification of the current task as well as the number classes of tasks
that the hardware supports.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>