x86/sgx: Introduce SGX notifier
KVM guest with SGX enabled should be paused to prevent from using
SGX memory while SGX CPUSVN update procedure is running, for any
active enclave in guest might fail the update process.

However, the CPUSVN update process is only running on host, which
guest is not aware of. It requires to provide a way to let host
notify KVM to pause guest. That is why SGX notifier is introduced.

SGX notifier is defined on host side and a notifier list is
maintained. Once SGX CPUSVN happens, host can tell all the
registered notifiers to take behavior.

Signed-off-by: Cathy Zhang <cathy.zhang@intel.com>
Signed-off-by: Qian Wen <qian.wen@linux.intel.com>