x86/tdx: Enable PLATFORM_VIRTIO_RESTRICTED_MEM_ACCESS for TDX
In virtio the host decides whether the guest uses the DMA
API or not using the strangely named VIRTIO_F_ACCESS_PLATFORM
bit (which really indicates whether the DMA API is used or not)

For hardened virtio on TDX, it is expected that swiotlb is
always used, which requires using the DMA API.  While IO wouldn't
really work without the swiotlb, it might be possible that an
attacker forces swiotlbless IO to manipulate memory in the guest.

So TDX has requirement to force the DMA API (which then forces
swiotlb), but without relying on the host.

Enable PLATFORM_VIRTIO_RESTRICTED_MEM_ACCESS platform feature.

Signed-off-by: Andi Kleen <ak@linux.intel.com>
Signed-off-by: Alexander Shishkin <alexander.shishkin@linux.intel.com>