idle: intel_idle: introduce debugfs I/F
The target_residency and exit_latency parameters are used by the cpuidle
framework to choose an idle state for a certain period of idle time.

Different values may bring different idle patterns for the same
workload, and results in power/performance difference.

In some cases, we need to evaluate the difference brought by the
different target_residency and exit_latency values, and previously, this
is done by modifying the internal table and rebuild the kernel.
This is not convenient, especially that some people who care about the
power/performance may not have experience of building a customized
kernel.

This debugfs I/F provides the capability of changing the idle state
parameters at runtime, without rebuilding the kernel.

Signed-off-by: Zhang Rui <rui.zhang@intel.com>