dmaengine: idxd: process user page faults for completion record
DSA supports page fault handling through PRS. However, the DMA engine
that's processing the descriptor is blocked until the PRS response is
received. Other workqueues sharing the engine are also blocked.
Page fault handing by the driver with PRS disabled can be used to
mitigate the stalling.

With PRS disabled while ATS remain enabled, DSA handles page faults on
a completion record by reporting an event in the event log. In this
instance, the descriptor is completed and the event log contains the
completion record adress and the contents of the completion record. Add
support to the event log handling code to fault in the completion record
and copy the content of the completion record to user memory.

A bitmap is introduced to keep track of discarded eventlog entries. When
the user process initiates ->release() of the char device, it no longer is
interested in any remaing event log entries tied to the relevant wq and
pasid. The driver will mark the event log entry index in the bitmap. Upon
encountering the entries during processing, the eventlog handler will just
clear the bitmap bit and skip the entry rather than attempt to process the
eventlog entry.

Signed-off-by: Dave Jiang <dave.jiang@intel.com>