perf annotate: Update the LBR event occurrences
The LBR event logging can indicate the precise event occurrences in LBRs,
which is useful for user to understand what causes long latency in program
(what events cause the long latency).

For each "LBR block", user would know that the selected counter(s)
incremented 0, 1, 2, or more times.

The max event occurrences will be exported through sysfs by kernel. But
at this moment, we just use a macro for max event occurrences definition.

We use event_occurs[] to sum the event occurrences per event and then we
will report the computed average number in next patch. But if the max event
occurrences is anywhere, it means we may lost some occurrences, so record
the status in max_occurs[]. In next patch, we will report '+' after the
average number to indicate we may lost track essentially.

Signed-off-by: Jin Yao <yao.jin@linux.intel.com>
Signed-off-by: Zhengjun Xing <zhengjun.xing@linux.intel.com>