thermal: intel: hfi: Accommodate systems with irregular thermal interrupts
There exist systems in the field (e.g., Lakefield) in which not all CPUs.
get the package level thermal interrupt. Furthermore, it is possible that
the CPUs that do receive the thermal interrupt are not ready to handle it
because the thermal vector in the local APIC has not been enabled. In such
cases it is possible to miss an HFI update it CPUs that are ready do not
receive the thermal interrupt and CPUs that are not ready do.

Work around this issue by checking the status of the HFI thermal interrupt
each time a CPU comes online. If it finds that the status bit is set, it
means that a missed interrupt happened. In such cases, consume the update
and queue work to process it.

Cc: Andi Kleen <ak@linux.intel.com>
Cc: Aubrey Li <aubrey.li@linux.intel.com>
Cc: Len Brown <len.brown@intel.com>
Cc: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>
Cc: Tim Chen <tim.c.chen@linux.intel.com>
Cc: "Ravi V. Shankar" <ravi.v.shankar@intel.com>
Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>