dmaengine: idxd: add support for inter-pasid table create window ioctl
Add ioctl function that creates an inter-domain pasid window for operation.
The "owner" is defined as the application that is opening a memory window
to allow another process to DMA into the owner's region. It sends the
create_window ioctl in order to create an IDPT entry. The owner can specify
whether the entry type should be type 0 (single submitter) or type 1
(multiple submitters). The driver will return a file handle to this entry.
The IDPT entry at this point has the valid bit set to 0. The ioctl also
returns a 16bit handle to the IDPT entry. This handle is used as a field in
the "window update" descriptor usable by the owner to update the window
parameters.

When the submitter is done with operations, it can close the IDPT entry
file handle. This will trigger the draining of the submitter pasid.

When the owner is done it will also close its IDPT entry file handle. This
effectively deletes the window and removes the IDPT entry.

Signed-off-by: Dave Jiang <dave.jiang@intel.com>