dmaengine: idxd: add support for idpt bitmap setup
The IDPT bitmap is used for supporting multi-submitter mode of IDPT
operation. The bitmap is a permission table to verify which PASIDs are
allowed to perform DMA operation for the memory window defined by the IDPT
entry. A pasid table requires holding 2^20 bits. If a device has thousands
of IDPT entries and a platform having multiple devices, a lot of memory
would be wasted for these bitmaps. By utilizing the device's capability of
accessing kernel virtual address (KVA), a KVA sparse map can be created
based on which pasids are actually being used. The device will attempt to
access the bitmap based on the submitter's pasid. The memory region that
does not have physical page backing will be read as 0 by the device.

A 2 level bitmap is created in this situation. The first level bitmap will
keep track of which pages are valid. And the second level bitmap is a KVA
sparsemap that only has page where the first level bitmap has a set bit.
The KVA sparsemap is created by using the get_vm_area() call for 128KB.
When a phyiscal page is needed, alloc_page() is called to get a page that
is inserted by vmap_pages_range() to the appropriate address range.

Signed-off-by: Dave Jiang <dave.jiang@intel.com>