x86/tdx: Implement ioremap_driver_hardened for x86
Implement ioremap_driver_hardened for x86. In TDX most memory is encrypted,
but some memory that is used to communicate with the host must
be declared shared with special page table attributes and a
special hypercall. Previously all ioremaped memory was declared
shared, but this leads to various BIOS tables and other private
state being shared.

Add a new ioremap_driver_hardened that shared the memory, and
make everything not shared.

This function is not intended to be used directly by drivers.

Instead it will be call by pci_iomap and the PCI subsystem
based on the authorized attribute of the device.

Signed-off-by: Andi Kleen <ak@linux.intel.com>
Signed-off-by: Kuppuswamy Sathyanarayanan <sathyanarayanan.kuppuswamy@linux.intel.com>