sched/fair: Identify two tasks of different classes to swap
A non-idle CPU doing migrate_misfit_task_class type of migration needs to
exchange a task from its runqueue with a tasks of the busiest runqueue.
Hence, but the busiest and the destination CPUs need to be stopped for the
task swap to proceed. Implement migrate_swap_stop2() that the CPU stopper
executes on either CPU to perform the live exchange.

When exchanging tasks, only consider tasks that can run on both CPUs and
that are classified.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>