platform/x86/intel/tpmi: Read feature control status
Some of the PM features can be locked or disabled. In that case, there
is no need to create a device for such feature.

This status is read via a mailbox. There is one TPMI ID which provides
base address for interface and data register for mail box operation.

This interface is called only during enumeration of TPMI devices and
via debugfs.

This patch is not tested yet. Need to wait for some HW support.

Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>