thermal: intel: hfi: Introduce Hardware Feedback Interface classes
On Intel hybrid parts, CPUs have asymmetric performance properties.
However, such asymmetry depends not only on the type of CPU but also
on the type of instructions a task executes. This means that different
types of tasks will have different instructions-per-cycle (IPC) ratios.

The Enhanced Hardware Feedback (EHFI) interface provides classification
of tasks based on the instructions they execute. In other words, hardware
can identify specific types of tasks with higher or lower IPC ratios.

Tasks are classified into several classes. Hardware defines the number
of supported classes.

Class 0 denotes tasks for which hardware did not find a suitable class.
The IPC ratio of class 0 is the average of all classes. Thus, use to decide
when to inject CPU idleness.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>