platform/x86/intel: TPMI domain id and CPU mapping
TPMI power domains include a group of CPUs. Several power management
settings in this case applicable to this group of CPUs not the whole
CPU package. So provide APIs to get power domain id for a CPU and get
mask of CPUs in a power domain.

Hardware punit uses different CPU numbering which is not based on APIC
numbering. But core masks are in punit numbering schemes, so provide an
API to map from each numbering scheme.

One each CPU online MSR 0x54 is used to read the mapping and stores in
a per cpu array. Also create a hash to fast search punit core id to
get Linux CPU number.

Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>