dmaengine: idxd: add wq driver name support for accel-config user tool
With the possibility of multiple wq drivers that can be bound to the wq,
the user config tool accel-config needs a way to know which wq driver to
bind to the wq. Introduce per wq driver_name sysfs attribute where the user
can indicate the driver to be bound to the wq. This allows accel-config to
just bind to the driver using wq->driver_name.

Signed-off-by: Dave Jiang <dave.jiang@intel.com>
Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>

 Conflicts:
	Documentation/ABI/stable/sysfs-driver-dma-idxd
	drivers/dma/idxd/cdev.c
	drivers/dma/idxd/dma.c
	drivers/dma/idxd/idxd.h
	drivers/dma/idxd/sysfs.c