irqdomain/msi: Provide msi_alloc/free_store() callbacks
For devices which don't have a standard storage for MSI messages
like the upcoming IMS (Interrupt Message Store) it's required to allocate
storage space before allocating interrupts and after freeing them.

This could be achieved with the existing callbacks, but that would
be awkward because they operate on msi_alloc_info_t which is not
uniform across architectures. Also these callbacks are invoked per interrupt
but the allocation might have bulk requirements depending on the device.

As such devices can operate on different architectures it is simpler
to have separate callbacks which operate on struct device. The
resulting storage information has to be stored in struct msi_desc so the
underlying irq chip implementation can retrieve it for the relevant operations.

Signed-off-by: Megha Dey <megha.dey@intel.com>