platform/x86: ISST: Add support for MSR 0x54
To map Linux CPU numbering scheme to hardware CPU numbering scheme
MSR 0x53 is getting used. But for new generation of CPUs, this MSR
is not valid. Since this is model specific MSR, this is possible.

A new MSR 0x54 is defined. Use this MSR and convert the IOCTL format
to match existing MSR 0x53, in this case user space don't need to
be aware of this change.

Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>

platform/x86: ISST: For user space use MSR 0x53 format

To avoid tool update and till we have support for MSR 54 widely just
reuse current MSR 0x53 format.

Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>